+++ app/models/bank_transaction.rb
@@ -19,6 +19,8 @@ class BankTransaction < ApplicationRecord
       .group_by(&:parent_transaction_type)
   end
 
+  scope :most_recent_first, -> { order(happened_at: :desc, created_at: :desc) }
+
   enum operation: {
     credit: 'credit'.freeze,
     debit: 'debit'.freeze
